requirements.txt
colorama==0.4.6
python-whois==0.9.5

